apply latest changes 2024-12-23

+++ MasterErp.Entities/DTOs/HR/EmployeeSalaryDto.cs
@@ -17,37 +17,37 @@ namespace MasterErp.Entities.DTOs.HR
         public DateTime RequestDate { get; set; }
         public bool? IsPaid { get; set; }
         public bool? IsGossi { get; set; }
-        public int? BasicSalary { get; set; }
-        public int? ExtraSalary { get; set; }
-        public int? WorkNature { get; set; }
-        public int? Transportation { get; set; }
-        public int? HousingAllowance { get; set; }
-        public int? MobileAllowance { get; set; }
-        public int? MealAllowance { get; set; }
-        public int? OverTime { get; set; }
-        public int? Other { get; set; }
-        public int? Other2 { get; set; }
-        public int? GrossSalary { get; set; }
+        public double? BasicSalary { get; set; }
+        public double? ExtraSalary { get; set; }
+        public double? WorkNature { get; set; }
+        public double? Transportation { get; set; }
+        public double? HousingAllowance { get; set; }
+        public double? MobileAllowance { get; set; }
+        public double? MealAllowance { get; set; }
+        public double? OverTime { get; set; }
+        public double? Other { get; set; }
+        public double? Other2 { get; set; }
+        public double? GrossSalary { get; set; }
         public int? DifferenceDays { get; set; }
-        public int? DifferenceCosteDays { get; set; }
-        public int? TotalGrossSalary { get; set; }
-        public int? Gosi { get; set; }
-        public int? Absence { get; set; }
-        public int? Penalty { get; set; }
-        public int? Loan { get; set; }
-        public int? SickCost { get; set; }
-        public int? TransportDeduct { get; set; }
-        public int? HomeDeduct { get; set; }
-        public int? OtherDeduct { get; set; }
-        public int? TotalDeduct { get; set; }
-        public int? NetSalary { get; set; }
+        public double? DifferenceCosteDays { get; set; }
+        public double? TotalGrossSalary { get; set; }
+        public double? Gosi { get; set; }
+        public double? Absence { get; set; }
+        public double? Penalty { get; set; }
+        public double? Loan { get; set; }
+        public double? SickCost { get; set; }
+        public double? TransportDeduct { get; set; }
+        public double? HomeDeduct { get; set; }
+        public double? OtherDeduct { get; set; }
+        public double? TotalDeduct { get; set; }
+        public double? NetSalary { get; set; }
         public bool? HRManagerApprove { get; set; }
         public bool? FinanceManagerApprove { get; set; }
         public bool? AuditingApprove { get; set; }
         public bool? ExecutiveManagerApprove { get; set; }
         public bool? GeneralManagerApprove { get; set; }
-        public int? VisaPayment { get; set; }
-        public int? CashPayment { get; set; }
-        public int? TotalSalary { get; set; }
+        public double? VisaPayment { get; set; }
+        public double? CashPayment { get; set; }
+        public double? TotalSalary { get; set; }
     }
 }
